moving story related actions to their own useStoryApi file in provider to clean up apiProvider file, doing the same for fonts.
